Restructured compute functions and added bond_counts (molecules wip)
